server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.3.20
langchain-openai==0.3.8
pica-langchain==1.4.0
python-dotenv==1.0.0
mcp-python==0.1.0 
requests==2.32.3